requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-core>=0.2.10
chromadb>=0.5.5
sentence-transformers>=3.0.1
ollama>=0.3.0
llama-cpp-python>=0.2.84
graphviz>=0.20.3
python-dotenv>=1.0.0
tree-sitter>=0.21.3
tree-sitter-cpp>=0.20.0
rich>=13.7.0

